Implement the Ruby version
This was fairly tricky to figure out, because the Google Protobuf tools
are fairly unpleasant to use. But the community gem solved the problems.

It follows the design of the Python version to a degree, but is a little
more Ruby like. That said, it's not expected you'd just drop this in as
a solution, but rather work it into your own situation.